mgr/cephadm: move nfs and osd to services/
