+++ src/pybind/mgr/cephadm/services/osd.py
